Fixed LayerProperties Bug, Added MomentumSGD (Optimizer) and Xavier (Initializer)
Added the MomentumSGD optimizer and the Xavier weight intializer. LayerProperties members were getting propagated as one shared object instead of individual objects for each layer.  This caused the same optimizer object to be passed to every single layer in the network causing errors because they needed independent internal state. I fixed this by making deep copies of the individual attributes on initialization. I also made the layer properties themselves to be instance variables of the layers instead of class variables.