Add Data Augmentation To Network Class
Added the ability to pass a list of data augmentation functions to the network class. These functions will be used on each training sample before they are passed the network for prediction. The gradient list used for batching in the network class was also converted to a numpy array instead of a python list. The README was also updated to reflect the data augmentation changes.